+++ src/commands/snap/label.go
@@ -9,7 +9,7 @@ import (
 func Label(snap int, label string) func(modl *model.Model) error {
 	return func(modl *model.Model) error {
 		for i := 0; i < len(modl.Snaps); i++ {
-			if modl.Snaps[i].Id == uint32(snap) {
+			if modl.Snaps[i].Id == snap {
 				modl.Snaps[i].Label = label
 				println("Ok.")
 				return nil
